+++ src/main/java/um/tesoreria/core/hexagonal/guarani/alumnoGuarani/application/usecases/CheckAllPreuniversitarioWithoutChequeraUseCaseImpl.java
@@ -46,6 +46,7 @@ public class CheckAllPreuniversitarioWithoutChequeraUseCaseImpl implements Check
             } catch (ChequeraSerieException e) {
                 log.error(e.getMessage());
                 pendientes.add(AlumnoDeteccionRequest.builder()
+                        .alumno(encontrado.getAlumno())
                         .ubicacion(encontrado.getUbicacion())
                         .propuesta(encontrado.getPropuesta())
                         .nroDocumento(encontrado.getNroDocumento())