+++ src/main/java/um/tesoreria/core/hexagonal/guarani/alumnoGuarani/domain/model/PersonaGuarani.java
@@ -1,5 +1,6 @@
 package um.tesoreria.core.hexagonal.guarani.alumnoGuarani.domain.model;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 
 import java.time.LocalDate;
@@ -12,7 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 public class PersonaGuarani {
 
-    private Long personaId;
+    private Long persona;
     private String apellido;
     private String nombres;
     private String apellidoElegido;
@@ -20,9 +21,13 @@ public class PersonaGuarani {
     private String sexo;
     private Integer identidadGenero;
     private String identidadGeneroOtro;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private LocalDate fechaNacimiento;
     private String localidadNacimiento;
     private Integer nacionalidad;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private LocalDate fechaIngresoPais;
     private String paisOrigen;
     private Long documentoPrincipal;
@@ -30,6 +35,8 @@ public class PersonaGuarani {
     private List<ContactoGuarani> contactos;
     private String usuario;
     private String clave;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private LocalDate fechaVencimientoClave;
     private String autenticacion;
     private Integer bloqueado;